fix: align buffer string limits with node
